kata_agent: Provide hotplugged device the right rootfs path
In case rootfs comes from a block device, we have to mount it to
the expected rootfs path without the "rootfs" suffix as this is
going to be exposed when mounted.

For instance if you mount the device to /foo/bar, you will have
the path /foo/bar/rootfs being your full rootfs path inside the
VM, that is going to be bind mounted by libcontainer inside the
container as the root directory.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>